feat(shortcodes): add "action"-variable to <form>-tag
This is to be used for *custom* success pages, see https://docs.netlify.com/forms/setup/#success-messages.

I will not include a custom success page in the example for now, since - I think - the standard netlify success message should work fine for a lot of use cases, and I want to keep this template as simple as possible.